+++ CHANGELOG.md
@@ -22,12 +22,12 @@ The types of changes are:
 ### Added
 - Additional consent reporting calls from `fides-js` [#3845](https://github.com/ethyca/fides/pull/3845)
 - Additional consent reporting calls from privacy center [#3847](https://github.com/ethyca/fides/pull/3847)
-- Adds server side prefetching experience, and enhance ip detection for to server-side geolocation [#3895](https://github.com/ethyca/fides/pull/3895)
 - Access support for Recurly [#3595](https://github.com/ethyca/fides/pull/3595)
 
 ### Fixed
 - Fix datamap zoom for low system counts [#3835](https://github.com/ethyca/fides/pull/3835)
 - Fixed connector forms with external dataset reference fields [#3873](https://github.com/ethyca/fides/pull/3873)
+- Fix ability to make server side API calls from privacy-center [#3895](https://github.com/ethyca/fides/pull/3895)
 
 ### Added